feat: 모델 관계 설정 수정 업장 서비스 예외처리 수정

+++ 1controllers/offer.controller.js
@@ -7,11 +7,17 @@ class OfferController {
   createOffer = async (req, res, next) => {
     const { offerName, offerNumber, price } = req.body;
 
-    //const { companyId } = req.params;
+    const { companyId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
-    const { status, message } = await this.offerService.createOffer(offerName, offerNumber, price);
+    const { status, message } = await this.offerService.createOffer(
+      userId,
+      companyId,
+      offerName,
+      offerNumber,
+      price
+    );
     res.status(status).json(message);
   };
 
@@ -21,14 +27,17 @@ class OfferController {
 
     const { companyId, offerId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
     const { status, message } = await this.offerService.updateOffer(
       offerId,
+      companyId,
+      userId,
       offerName,
       offerNumber,
       price
     );
+
     res.status(status).json(message);
   };
 
@@ -36,9 +45,9 @@ class OfferController {
   deleteOffer = async (req, res, next) => {
     const { offerId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
-    const { status, message } = await this.offerService.destroyOffer(offerId);
+    const { status, message } = await this.offerService.destroyOffer(offerId, userId);
     res.status(status).json(message);
   };
